Change of approach
We changed the approach to run simulations: we generate a single gif per grid size and total number of steps, instead of 1 for each number of steps. This significantly reduces simulation time and makes it possible to actually finish this project on time.

Other changes include renaming the sh script to generate plots, changing all instances of heatmaps to frames so that it's more clear, added other graphs to be generated for more detailed report.